feat : 검색화면 리사이클러뷰 레이아웃 위치 변경, 검색결과 비건 타입 아이콘 바인딩 어댑터 설정
