+++ app/src/main/java/com/android/aviro/domain/entity/SearchedRestaurantItem.kt
@@ -1,0 +1,3 @@
+package com.android.aviro.domain.entity
+
+data class SearchedRestaurantItem()
