+++ app/src/main/java/com/android/aviro/presentation/search/SortingLocEntity.kt
@@ -1,0 +1,4 @@
+package com.android.aviro.presentation.search
+
+class SortingLocEntity {
+}
