+++ app/src/main/res/drawable/ic_dropdown.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M11.998,17L18.998,9H4.998L11.998,17Z"
+      android:fillColor="#6B7581"/>
+</vector>
