+++ app/src/main/res/layout/activity_search.xml
@@ -202,16 +202,17 @@
 
         </LinearLayout>
 
-        <!-- 이전 검색어  -->
+        <!-- 최근 검색어  -->
         <LinearLayout
             android:id="@+id/includeLayout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
             app:visibilityChanged="@{(viewmodel.isPreSearched &amp;&amp; !viewmodel.isSearching)}">
+
             <LinearLayout
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
+                android:layout_height="54dp"
                 android:orientation="horizontal">
 
         <TextView
@@ -252,6 +253,68 @@
                 android:textSize="14sp" />
         </LinearLayout>
             </LinearLayout>
+
+            <!--최근 검색어 리스트뷰-->
+
+        </LinearLayout>
+
+    <!-- 정렬방식  -->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="54dp"
+        android:orientation="horizontal"
+        app:visibilityChanged="@{viewmodel.isSearching}">
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="54dp"
+            android:paddingStart="16dp"
+            android:paddingEnd="16dp"
+            android:paddingTop="15dp"
+            android:paddingBottom="15dp"
+            android:orientation="horizontal">
+            <LinearLayout
+                android:id="@+id/locationSortBtn"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal"
+                android:onClick="@{() -> viewmodel.onClickLocSort(locationSortBtn)}">
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@{viewmodel.SortLocText}"
+                android:fontFamily="@font/pretendard_medium"
+                android:textColor="@color/Gray3"
+                android:textSize="16sp"/>
+            <LinearLayout
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                android:background="@drawable/ic_dropdown"
+                android:orientation="horizontal"/>
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@+id/accuracySortBtn"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal"
+                android:onClick="@{() -> viewmodel.onClickAccSort(accuracySortBtn)}">
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="10dp"
+                android:text="@{viewmodel.SortAccText}"
+                android:fontFamily="@font/pretendard_medium"
+                android:textColor="@color/Gray3"
+                android:textSize="16sp"/>
+            <LinearLayout
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                android:background="@drawable/ic_dropdown"
+                android:orientation="horizontal"/>
+            </LinearLayout>
+        </LinearLayout>
+
+
 
         </LinearLayout>
